Refactored application tests to use EventStoreDB
