+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/immutable/functions/shoppingCart.ts
@@ -1,4 +1,4 @@
-import { merge } from './core';
+import { merge } from '../../tools/utils';
 
 export interface ProductItem {
   productId: string;
