+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/oop/aggregate/aggregate.ts
@@ -1,0 +1,19 @@
+import { Event } from '../../tools/events';
+
+export abstract class Aggregate<E extends Event> {
+  #uncommitedEvents: E[] = [];
+
+  abstract evolve(event: E): void;
+
+  protected enqueue = (event: E) => {
+    this.#uncommitedEvents = [...this.#uncommitedEvents, event];
+  };
+
+  dequeueUncommitedEvents = (): E[] => {
+    const events = this.#uncommitedEvents;
+
+    this.#uncommitedEvents = [];
+
+    return events;
+  };
+}
