+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/oop/aggregate/shoppingCart.ts
@@ -1,6 +1,6 @@
 import { ShoppingCartErrors } from './businessLogic';
-import { Aggregate } from './core';
-import { Event } from './core';
+import { Aggregate } from './aggregate';
+import { Event } from '../../tools/events';
 
 export interface ProductItem {
   productId: string;
