+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/oop/aggregate_returning_events/applicationLogic.exercise.test.ts
@@ -1,28 +1,34 @@
 import { v4 as uuid } from 'uuid';
+import { getEventStoreDBTestClient } from '#core/testing/eventStoreDB';
+import { EventStoreDBClient } from '@eventstore/db-client';
+import { getEventStore } from '../../tools/eventStore';
 import {
-  ShoppingCartErrors,
-  openShoppingCart,
-  addProductItemToShoppingCart,
-  removeProductItemFromShoppingCart,
-  confirmShoppingCart,
-  cancelShoppingCart,
-  OpenShoppingCart,
   AddProductItemToShoppingCart,
-  RemoveProductItemFromShoppingCart,
-  ConfirmShoppingCart,
   CancelShoppingCart,
+  ConfirmShoppingCart,
+  OpenShoppingCart,
+  RemoveProductItemFromShoppingCart,
+  ShoppingCartErrors,
 } from './businessLogic';
 import {
   PricedProductItem,
+  ShoppingCart,
   ShoppingCartEvent,
   ShoppingCartStatus,
   getShoppingCart,
 } from './shoppingCart';
-import { getEventStore } from './core';
 
 describe('Business logic', () => {
-  it('Should handle commands correctly', () => {
-    const eventStore = getEventStore();
+  let eventStoreDB: EventStoreDBClient;
+
+  beforeAll(async () => {
+    eventStoreDB = await getEventStoreDBTestClient();
+  });
+
+  afterAll(() => eventStoreDB.dispose());
+
+  it('Should handle commands correctly', async () => {
+    const eventStore = getEventStore(eventStoreDB);
     const shoppingCartId = uuid();
 
     const clientId = uuid();
@@ -50,15 +56,19 @@ describe('Business logic', () => {
       unitPrice: 5,
     };
 
-    let result: ShoppingCartEvent[] = [];
-
     // Open
     const open: OpenShoppingCart = {
       type: 'OpenShoppingCart',
       data: { shoppingCartId, clientId, now: openedAt },
     };
-    result = [openShoppingCart(open)];
-    eventStore.appendToStream(shoppingCartId, ...result);
+    await eventStore.appendToStream(
+      shoppingCartId,
+      ShoppingCart.open(
+        open.data.shoppingCartId,
+        open.data.clientId,
+        open.data.now,
+      ),
+    );
 
     // Add Two Pair of Shoes
     const addTwoPairsOfShoes: AddProductItemToShoppingCart = {
@@ -66,10 +76,14 @@ describe('Business logic', () => {
       data: { shoppingCartId, productItem: twoPairsOfShoes },
     };
 
-    let state = getShoppingCart(eventStore.readStream(shoppingCartId));
-    result = [addProductItemToShoppingCart(addTwoPairsOfShoes, state)];
+    let shoppingCart = getShoppingCart(
+      await eventStore.readStream(shoppingCartId),
+    );
 
-    eventStore.appendToStream(shoppingCartId, ...result);
+    await eventStore.appendToStream(
+      shoppingCartId,
+      shoppingCart.addProductItem(addTwoPairsOfShoes.data.productItem),
+    );
 
     // Add T-Shirt
     const addTShirt: AddProductItemToShoppingCart = {
@@ -77,9 +91,12 @@ describe('Business logic', () => {
       data: { shoppingCartId, productItem: tShirt },
     };
 
-    state = getShoppingCart(eventStore.readStream(shoppingCartId));
-    result = [addProductItemToShoppingCart(addTShirt, state)];
-    eventStore.appendToStream(shoppingCartId, ...result);
+    shoppingCart = getShoppingCart(await eventStore.readStream(shoppingCartId));
+
+    await eventStore.appendToStream(
+      shoppingCartId,
+      shoppingCart.addProductItem(addTShirt.data.productItem),
+    );
 
     // Remove pair of shoes
     const removePairOfShoes: RemoveProductItemFromShoppingCart = {
@@ -87,9 +104,11 @@ describe('Business logic', () => {
       data: { shoppingCartId, productItem: pairOfShoes },
     };
 
-    state = getShoppingCart(eventStore.readStream(shoppingCartId));
-    result = [removeProductItemFromShoppingCart(removePairOfShoes, state)];
-    eventStore.appendToStream(shoppingCartId, ...result);
+    shoppingCart = getShoppingCart(await eventStore.readStream(shoppingCartId));
+    await eventStore.appendToStream(
+      shoppingCartId,
+      shoppingCart.removeProductItem(removePairOfShoes.data.productItem),
+    );
 
     // Confirm
     const confirm: ConfirmShoppingCart = {
@@ -97,24 +116,31 @@ describe('Business logic', () => {
       data: { shoppingCartId, now: confirmedAt },
     };
 
-    state = getShoppingCart(eventStore.readStream(shoppingCartId));
-    result = [confirmShoppingCart(confirm, state)];
-    eventStore.appendToStream(shoppingCartId, ...result);
+    shoppingCart = getShoppingCart(await eventStore.readStream(shoppingCartId));
+    await eventStore.appendToStream(
+      shoppingCartId,
+      shoppingCart.confirm(confirm.data.now),
+    );
 
     // Try Cancel
     const cancel: CancelShoppingCart = {
       type: 'CancelShoppingCart',
       data: { shoppingCartId, now: canceledAt },
     };
-    const onCancel = () => {
-      state = getShoppingCart(eventStore.readStream(shoppingCartId));
-      result = [cancelShoppingCart(cancel, state)];
-      eventStore.appendToStream(shoppingCartId, ...result);
+    const onCancel = async () => {
+      shoppingCart = getShoppingCart(
+        await eventStore.readStream(shoppingCartId),
+      );
+      await eventStore.appendToStream(
+        shoppingCartId,
+        shoppingCart.cancel(cancel.data.now),
+      );
     };
 
     expect(onCancel).toThrow(ShoppingCartErrors.CART_IS_ALREADY_CLOSED);
 
-    const events = eventStore.readStream<ShoppingCartEvent>(shoppingCartId);
+    const events =
+      await eventStore.readStream<ShoppingCartEvent>(shoppingCartId);
 
     expect(events).toEqual([
       {
@@ -160,14 +186,24 @@ describe('Business logic', () => {
       // },
     ]);
 
-    const shoppingCart = getShoppingCart(events);
+    shoppingCart = getShoppingCart(events);
 
-    expect(shoppingCart).toStrictEqual({
-      id: shoppingCartId,
+    expect(shoppingCart).toBeInstanceOf(ShoppingCart);
+    const actual = {
+      shoppingCartId: shoppingCart.id,
+      clientId: shoppingCart.clientId,
+      status: shoppingCart.status,
+      openedAt: shoppingCart.openedAt,
+      productItems: shoppingCart.productItems,
+      confirmedAt: shoppingCart.confirmedAt,
+    };
+
+    expect(actual).toEqual({
+      shoppingCartId,
       clientId,
       status: ShoppingCartStatus.Confirmed,
-      productItems: [pairOfShoes, tShirt],
       openedAt,
+      productItems: [pairOfShoes, tShirt],
       confirmedAt,
     });
   });
