+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/oop/aggregate_returning_events/shoppingCart.ts
@@ -1,5 +1,5 @@
 import { ShoppingCartErrors } from './businessLogic';
-import { Event } from './core';
+import { Event } from '../../tools/events';
 
 export interface ProductItem {
   productId: string;
