+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/tools/commands.ts
@@ -1,0 +1,7 @@
+export type Command<
+  CommandType extends string = string,
+  CommandData extends Record<string, unknown> = Record<string, unknown>,
+> = Readonly<{
+  type: Readonly<CommandType>;
+  data: Readonly<CommandData>;
+}>;
