+++ workshops/introduction_to_event_sourcing/src/06_application_logic_eventstoredb/tools/events.ts
@@ -1,0 +1,7 @@
+export type Event<
+  EventType extends string = string,
+  EventData extends Record<string, unknown> = Record<string, unknown>,
+> = Readonly<{
+  type: Readonly<EventType>;
+  data: Readonly<EventData>;
+}>;
