+++ workshops/introduction_to_event_sourcing/src/07_optimistic_concurrency_eventstoredb/oop/solution1/businessLogic.ts
@@ -13,7 +13,7 @@ import {
 import { Event } from './optimisticConcurrency.exercise.solved';
 
 export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: Event[] = [];
+  #uncommitedEvents: E[] = [];
 
   abstract evolve(event: E): void;
