+++ workshops/introduction_to_event_sourcing/src/solved/05_business_logic/oop/solution1/businessLogic.ts
@@ -6,7 +6,7 @@ import {
 import { Event, EventStore } from './businessLogic.solved.test';
 
 export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: Event[] = [];
+  #uncommitedEvents: E[] = [];
 
   abstract evolve(event: E): void;
