+++ workshops/introduction_to_event_sourcing/src/solved/05_command_definition/oop/solution1/core.ts
@@ -7,7 +7,7 @@ export type Event<
 }>;
 
 export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: Event[] = [];
+  #uncommitedEvents: E[] = [];
 
   abstract evolve(event: E): void;
