+++ workshops/introduction_to_event_sourcing/src/solved/06_business_logic_eventstoredb/oop/solution1/businessLogic.ts
@@ -11,7 +11,7 @@ import {
 import { Event } from './businessLogic.solved.test';
 
 export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: Event[] = [];
+  #uncommitedEvents: E[] = [];
 
   abstract evolve(event: E): void;
