+++ workshops/introduction_to_event_sourcing/src/solved/07_optimistic_concurrency_eventstoredb/implicitExpectedRevision/oop/solution1/businessLogic.ts
@@ -14,7 +14,7 @@ import {
 import { Event } from './optimisticConcurrency.exercise.test';
 
 export abstract class Aggregate<E extends Event> {
-  #uncommitedEvents: Event[] = [];
+  #uncommitedEvents: E[] = [];
 
   abstract evolve(event: E): void;
